+++ src/ext/BalExtension/wixstdba/WixStandardBootstrapperApplication.cpp
@@ -186,6 +186,16 @@ static THEME_ASSIGN_CONTROL_ID vrgInitControls[] = {
     { WIXSTDBA_CONTROL_FAILURE_CANCEL_BUTTON, L"FailureCloseButton" },
 };
 
+typedef struct _WIXSTDBA_PREREQ_PACKAGE
+{
+    LPWSTR sczPackageId;
+    BAL_INFO_PACKAGE* pPackage;
+    BOOL fAlwaysInstall;
+    BOOL fWasAlreadyInstalled;
+    BOOL fPlannedToBeInstalled;
+    BOOL fSuccessfullyInstalled;
+} WIXSTDBA_PREREQ_PACKAGE;
+
 class CWixStandardBootstrapperApplication : public CBalBaseBootstrapperApplication
 {
 public: // IBootstrapperApplication
@@ -225,7 +235,7 @@ public: // IBootstrapperApplication
                 nResult = IDRESTART;
             }
 
-            if (m_sczPrereqPackage)
+            if (m_fPrereq)
             {
                 BalLog(BOOTSTRAPPER_LOG_LEVEL_STANDARD, m_fAllowRestart ? "The prerequisites scheduled a restart. The bootstrapper application will be reloaded after the computer is restarted."
                                                                         : "A restart is required by the prerequisites but the user delayed it. The bootstrapper application will be reloaded after the computer is restarted.");
@@ -261,7 +271,7 @@ public: // IBootstrapperApplication
         BalInfoAddRelatedBundleAsPackage(&m_Bundle.packages, wzBundleId, relationType, fPerMachine);
 
         // If we're not doing a prerequisite install, remember when our bundle would cause a downgrade.
-        if (!m_sczPrereqPackage && BOOTSTRAPPER_RELATED_OPERATION_DOWNGRADE == operation)
+        if (!m_fPrereq && BOOTSTRAPPER_RELATED_OPERATION_DOWNGRADE == operation)
         {
             m_fDowngrading = TRUE;
         }
@@ -276,11 +286,12 @@ public: // IBootstrapperApplication
         __in BOOTSTRAPPER_PACKAGE_STATE state
         )
     {
-        // If the prerequisite package is already installed, remember that.
-        if (m_sczPrereqPackage && BOOTSTRAPPER_PACKAGE_STATE_PRESENT == state &&
-            CSTR_EQUAL == ::CompareStringW(LOCALE_NEUTRAL, 0, wzPackageId, -1, m_sczPrereqPackage, -1))
+        WIXSTDBA_PREREQ_PACKAGE* pPrereqPackage = NULL;
+        HRESULT hr = GetPrereqPackage(wzPackageId, &pPrereqPackage);
+        if (SUCCEEDED(hr) && BOOTSTRAPPER_PACKAGE_STATE_PRESENT == state)
         {
-            m_fPrereqAlreadyInstalled = TRUE;
+            // If the prerequisite package is already installed, remember that.
+            pPrereqPackage->fWasAlreadyInstalled = TRUE;
         }
     }
 
@@ -298,6 +309,17 @@ public: // IBootstrapperApplication
         if (SUCCEEDED(hrStatus))
         {
             hrStatus = EvaluateConditions();
+            
+            m_fPrereqAlreadyInstalled = TRUE;
+
+            // At this point we have to assume that all prerequisite packages need to be installed, so set to false if any of them aren't installed.
+            for (UINT i = 0; i < m_cPrereqPackages; ++i)
+            {
+                if (m_rgPrereqPackages[i].sczPackageId && !m_rgPrereqPackages[i].fWasAlreadyInstalled)
+                {
+                    m_fPrereqAlreadyInstalled = FALSE;
+                }
+            }
         }
 
         SetState(WIXSTDBA_STATE_DETECTED, hrStatus);
@@ -319,8 +341,8 @@ public: // IBootstrapperApplication
         __inout_z BOOTSTRAPPER_REQUEST_STATE* pRequestedState
         )
     {
-        // If we're only installing prereq, do not touch related bundles.
-        if (m_sczPrereqPackage)
+        // If we're only installing prerequisites, do not touch related bundles.
+        if (m_fPrereq)
         {
             *pRequestedState = BOOTSTRAPPER_REQUEST_STATE_NONE;
         }
@@ -334,20 +356,49 @@ public: // IBootstrapperApplication
         __inout BOOTSTRAPPER_REQUEST_STATE *pRequestState
         )
     {
-        // If we're planning to install a pre-req, install it. The pre-req needs to be installed
+        HRESULT hr = S_OK;
+        WIXSTDBA_PREREQ_PACKAGE* pPrereqPackage = NULL;
+
+        // If we're planning to install a prerequisite, install it. The prerequisite needs to be installed
         // in all cases (even uninstall!) so the BA can load next.
-        if (m_sczPrereqPackage)
+        if (m_fPrereq)
         {
-            if (CSTR_EQUAL == ::CompareStringW(LOCALE_NEUTRAL, 0, wzPackageId, -1, m_sczPrereqPackage, -1))
+            // Only install prerequisite packages, and check the InstallCondition on prerequisite support packages.
+            BOOL fInstall = FALSE;
+            hr = GetPrereqPackage(wzPackageId, &pPrereqPackage);
+            if (SUCCEEDED(hr) && pPrereqPackage->pPackage)
+            {
+                if (pPrereqPackage->fAlwaysInstall)
+                {
+                    fInstall = TRUE;
+                }
+                else if(pPrereqPackage->pPackage->sczInstallCondition && *pPrereqPackage->pPackage->sczInstallCondition)
+                {
+                    hr = m_pEngine->EvaluateCondition(pPrereqPackage->pPackage->sczInstallCondition, &fInstall);
+                    if (FAILED(hr))
+                    {
+                        fInstall = FALSE;
+                    }
+                }
+                else
+                {
+                    // If the InstallCondition is missing, then it should always be installed.
+                    fInstall = TRUE;
+                }
+
+                pPrereqPackage->fPlannedToBeInstalled = fInstall;
+            }
+
+            if (fInstall)
             {
                 *pRequestState = BOOTSTRAPPER_REQUEST_STATE_PRESENT;
             }
-            else // skip everything else.
+            else
             {
                 *pRequestState = BOOTSTRAPPER_REQUEST_STATE_NONE;
             }
         }
-        else if (m_sczAfterForcedRestartPackage) // after force restart skip packages until after the package that caused the restart.
+        else if (m_sczAfterForcedRestartPackage) // after force restart, skip packages until after the package that caused the restart.
         {
             // After restart we need to finish the dependency registration for our package so allow the package
             // to go present.
@@ -381,6 +432,17 @@ public: // IBootstrapperApplication
         {
             BalLog(BOOTSTRAPPER_LOG_LEVEL_STANDARD, "Running plan complete BA function");
             m_pBAFunction->OnPlanComplete();
+        }
+
+        m_fPrereqAlreadyInstalled = TRUE;
+
+        // Now that we've planned the packages, we can focus on the prerequisite packages that are supposed to be installed.
+        for (UINT i = 0; i < m_cPrereqPackages; ++i)
+        {
+            if (m_rgPrereqPackages[i].sczPackageId && !m_rgPrereqPackages[i].fWasAlreadyInstalled && m_rgPrereqPackages[i].fPlannedToBeInstalled)
+            {
+                m_fPrereqAlreadyInstalled = FALSE;
+            }
         }
 
         SetState(WIXSTDBA_STATE_PLANNED, hrStatus);
@@ -688,11 +750,13 @@ public: // IBootstrapperApplication
 
         int nResult = __super::OnExecutePackageComplete(wzPackageId, hrExitCode, restart, nRecommendation);
 
-        if (m_sczPrereqPackage && CSTR_EQUAL == ::CompareStringW(LOCALE_NEUTRAL, 0, wzPackageId, -1, m_sczPrereqPackage, -1))
+        WIXSTDBA_PREREQ_PACKAGE* pPrereqPackage = NULL;
+        HRESULT hr = GetPrereqPackage(wzPackageId, &pPrereqPackage);
+        if (SUCCEEDED(hr))
         {
-            m_fPrereqInstalled = SUCCEEDED(hrExitCode);
+            pPrereqPackage->fSuccessfullyInstalled = TRUE;
 
-            // If the pre-req required a restart (any restart) then do an immediate
+            // If the prerequisite required a restart (any restart) then do an immediate
             // restart to ensure that the bundle will get launched again post reboot.
             if (BOOTSTRAPPER_APPLY_RESTART_NONE != restart)
             {
@@ -782,6 +846,26 @@ public: // IBootstrapperApplication
         m_fRestartRequired = (BOOTSTRAPPER_APPLY_RESTART_NONE != restart && BOOTSTRAPPER_RESTART_NEVER < m_command.restart);
         // If a restart is required and we're not displaying a UI or we are not supposed to prompt for restart then allow the restart.
         m_fAllowRestart = m_fRestartRequired && (BOOTSTRAPPER_DISPLAY_FULL > m_command.display || BOOTSTRAPPER_RESTART_PROMPT < m_command.restart);
+
+        m_fPrereqInstalled = TRUE;
+        BOOL fInstalledAPackage = FALSE;
+
+        for (UINT i = 0; i < m_cPrereqPackages; ++i)
+        {
+            if (m_rgPrereqPackages[i].sczPackageId && m_rgPrereqPackages[i].fPlannedToBeInstalled)
+            {
+                if (m_rgPrereqPackages[i].fSuccessfullyInstalled)
+                {
+                    fInstalledAPackage = TRUE;
+                }
+                else
+                {
+                    m_fPrereqInstalled = FALSE;
+                }
+            }
+        }
+
+        m_fPrereqInstalled = m_fPrereqInstalled && fInstalledAPackage;
 
         // If we are showing UI, wait a beat before moving to the final screen.
         if (BOOTSTRAPPER_DISPLAY_NONE < m_command.display)
@@ -886,7 +970,7 @@ private: // privates
 
 
     //
-    // InitializeData - initializes all the package and prereq information.
+    // InitializeData - initializes all the package and prerequisite information.
     //
     HRESULT InitializeData()
     {
@@ -1148,6 +1232,11 @@ private: // privates
     {
         HRESULT hr = S_OK;
         IXMLDOMNode* pNode = NULL;
+        IXMLDOMNodeList* pNodes = NULL;
+        DWORD cNodes = 0;
+        LPWSTR scz = NULL;
+        WIXSTDBA_PREREQ_PACKAGE* pPrereqPackage = NULL;
+        BAL_INFO_PACKAGE* pPackage = NULL;
 
         hr = XmlSelectSingleNode(pixdManifest, L"/BootstrapperApplicationData/WixMbaPrereqInformation", &pNode);
         if (S_FALSE == hr)
@@ -1173,8 +1262,77 @@ private: // privates
         }
         BalExitOnFailure(hr, "Failed to get prerequisite license file.");
 
+        // get the list of prerequisite support packages
+        hr = XmlSelectNodes(pixdManifest, L"/BootstrapperApplicationData/MbaPrerequisiteSupportPackage", &pNodes);
+        if (S_FALSE == hr)
+        {
+            ExitFunction1(hr = S_OK);
+        }
+        ExitOnFailure(hr, "Failed to select prerequisite support package nodes.");
+
+        hr = pNodes->get_length((long*)&cNodes);
+        ExitOnFailure(hr, "Failed to get prerequisite support package node count.");
+
+        m_cPrereqPackages = cNodes + 1;
+        m_rgPrereqPackages = static_cast<WIXSTDBA_PREREQ_PACKAGE*>(MemAlloc(sizeof(WIXSTDBA_PREREQ_PACKAGE) * m_cPrereqPackages, TRUE));
+
+        hr = DictCreateWithEmbeddedKey(&m_shPrereqSupportPackages, m_cPrereqPackages, reinterpret_cast<void **>(&m_rgPrereqPackages), offsetof(WIXSTDBA_PREREQ_PACKAGE, sczPackageId), DICT_FLAG_NONE);
+        ExitOnFailure(hr, "Failed to create the prerequisite package dictionary.");
+
+        pPrereqPackage = m_rgPrereqPackages;
+        pPrereqPackage->sczPackageId = m_sczPrereqPackage;
+        pPrereqPackage->fAlwaysInstall = TRUE;
+        hr = BalInfoFindPackageById(&m_Bundle.packages, pPrereqPackage->sczPackageId, &pPrereqPackage->pPackage);
+        // Ignore error.
+        hr = DictAddValue(m_shPrereqSupportPackages, pPrereqPackage);
+        ExitOnFailure1(hr, "Failed to add \"%ls\" to the prerequisite package dictionary.", pPrereqPackage->sczPackageId);
+
+        for (DWORD i = 0; i < cNodes; ++i)
+        {
+            hr = XmlNextElement(pNodes, &pNode, NULL);
+            ExitOnFailure(hr, "Failed to get next node.");
+
+            // @PackageId
+            hr = XmlGetAttributeEx(pNode, L"PackageId", &scz);
+            ExitOnFailure(hr, "Failed to get @PackageId.");
+
+            hr = DictGetValue(m_shPrereqSupportPackages, scz, reinterpret_cast<void **>(&pPrereqPackage));
+            if (SUCCEEDED(hr))
+            {
+                if (CSTR_EQUAL == ::CompareStringW(LOCALE_NEUTRAL, 0, scz, -1, m_sczPrereqPackage, -1))
+                {
+                    pPrereqPackage->fAlwaysInstall = FALSE;
+                }
+                ReleaseNullObject(pNode);
+                continue;
+            }
+            else if (E_NOTFOUND != hr)
+            {
+                ExitOnFailure1(hr, "Failed to check if \"%ls\" was in the prerequisite package dictionary.", scz);
+            }
+
+            hr = BalInfoFindPackageById(&m_Bundle.packages, scz, &pPackage);
+            if (SUCCEEDED(hr))
+            {
+                pPrereqPackage = &m_rgPrereqPackages[i + 1];
+                pPrereqPackage->sczPackageId = pPackage->sczId;
+                pPrereqPackage->pPackage = pPackage;
+                hr = DictAddValue(m_shPrereqSupportPackages, pPrereqPackage);
+                ExitOnFailure1(hr, "Failed to add \"%ls\" to the prerequisite package dictionary.", pPrereqPackage->sczPackageId);
+            }
+            else
+            {
+                BalLogError(hr, "Failed to get info about \"%ls\" from BootstrapperApplicationData.", scz);
+            }
+
+            // prepare next iteration
+            ReleaseNullObject(pNode);
+        }
+
     LExit:
         ReleaseObject(pNode);
+        ReleaseObject(pNodes);
+        ReleaseStr(scz);
         return hr;
     }
 
@@ -1265,6 +1423,34 @@ private: // privates
 
     LExit:
         ReleaseObject(pNode);
+        return hr;
+    }
+
+    HRESULT GetPrereqPackage(
+        __in_z LPCWSTR wzPackageId,
+        __out WIXSTDBA_PREREQ_PACKAGE** ppPrereqPackage
+        )
+    {
+        HRESULT hr = E_NOTFOUND;
+        WIXSTDBA_PREREQ_PACKAGE* pPrereqPackage = NULL;
+
+        Assert(wzPackageId && *wzPackageId);
+        Assert(ppPrereqPackage);
+
+        if (m_shPrereqSupportPackages)
+        {
+            hr = DictGetValue(m_shPrereqSupportPackages, wzPackageId, reinterpret_cast<void **>(&pPrereqPackage));
+            if (E_NOTFOUND != hr)
+            {
+                ExitOnFailure(hr, "Failed to check the dictionary of prerequisite packages.");
+            }
+        }
+
+        if (pPrereqPackage)
+        {
+            *ppPrereqPackage = pPrereqPackage;
+        }
+    LExit:
         return hr;
     }
 
@@ -2695,6 +2881,9 @@ public:
         m_fShowVersion = FALSE;
 
         m_sdOverridableVariables = NULL;
+        m_shPrereqSupportPackages = NULL;
+        m_rgPrereqPackages = NULL;
+        m_cPrereqPackages = 0;
         m_pTaskbarList = NULL;
         m_uTaskbarButtonCreatedMessage = UINT_MAX;
         m_fTaskbarButtonOK = FALSE;
@@ -2723,6 +2912,8 @@ public:
 
         ReleaseObject(m_pTaskbarList);
         ReleaseDict(m_sdOverridableVariables);
+        ReleaseDict(m_shPrereqSupportPackages);
+        ReleaseMem(m_rgPrereqPackages);
         ReleaseStr(m_sczFailedMessage);
         ReleaseStr(m_sczConfirmCloseMessage);
         BalConditionsUninitialize(&m_Conditions);
@@ -2785,6 +2976,10 @@ private:
     BOOL m_fShowVersion;
 
     STRINGDICT_HANDLE m_sdOverridableVariables;
+
+    WIXSTDBA_PREREQ_PACKAGE* m_rgPrereqPackages;
+    UINT m_cPrereqPackages;
+    STRINGDICT_HANDLE m_shPrereqSupportPackages;
 
     BOOL m_fPrereq;
     LPWSTR m_sczPrereqPackage;